+++ examples/ros-diagnostics-msgs/src/App.tsx
@@ -2,11 +2,9 @@ import "./App.css";
 import { Table } from "./components/table";
 import useModuleDataListener from "./hooks/useDataModuleListener";
 import { useFetchDiagnostics } from "./hooks/useFetchDiagnostics";
-import { LoadingIndicator, useFormant } from "@formant/ui-sdk";
+import { useFormant } from "@formant/ui-sdk";
 import { dummyData } from "./util";
 import { IConfiguration } from "./types/types";
-import { useEffect, useMemo } from "react";
-import { data } from "./data";
 
 function App() {
   const context = useFormant();