+++ examples/ros-diagnostics-msgs/src/components/MessageTable/index.tsx
@@ -1,8 +1,8 @@
 import { MessageRow } from "../MessageRow";
-import { DiagnosticStatusMessage, IStatus } from "../../types/types";
+import { IStatus } from "../../types/types";
 import { FC } from "react";
-import styles from "./index.module.scss";
 import styled from "@emotion/styled";
+import { mediaQueries } from "../../styles/breakpoints";
 
 interface IMessageTable {
   messages: IStatus[];
@@ -16,30 +16,35 @@ export const MessageTable: FC<IMessageTable> = ({
   handleSetActive,
 }) => {
   return (
-    <div
-      className={styles["message-table"]}
+    <Container
       style={{
         minWidth: active === null ? "100%" : 500,
       }}
     >
-      {messages.map((_) => (
-        // <Row>
-        //   {_.name}
-        // </Row>
-        <MessageRow
-          key={_.name}
-          level={_.level}
-          active={active}
-          setActive={() => {
-            handleSetActive(_.name);
-          }}
-          message={_.name}
-        />
-      ))}
-    </div>
+      {messages
+        .sort((a, b) => a.name.localeCompare(b.name))
+        .map((_) => (
+          <MessageRow
+            key={_.name}
+            level={_.level}
+            active={active}
+            setActive={() => {
+              handleSetActive(_.name);
+            }}
+            message={_.name}
+          />
+        ))}
+    </Container>
   );
 };
 
-const Row = styled.div`
-  height: 40px;
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  background-color: #282f45;
+  min-width: 500px;
+  overflow: auto;
+  ${mediaQueries.small} {
+    width: 100vw;
+  }
 `;