+++ packages/data-sdk/src/api/getAllEventTriggerGroup.ts
@@ -1,0 +1,18 @@
+import { IEventTriggerGroup } from "../model/IEventTriggerGroup";
+import { Authentication } from "../Authentication";
+import { FORMANT_API_URL } from "../config";
+
+export async function getAllEventTriggerGroup(): Promise<IEventTriggerGroup[]> {
+  if (!Authentication.token) {
+    throw new Error("Not authenticated");
+  }
+  const data = await fetch(`${FORMANT_API_URL}/v1/admin/event-trigger-groups`, {
+    method: "GET",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: "Bearer " + Authentication.token,
+    },
+  });
+
+  return (await data.json()).items as IEventTriggerGroup[];
+}