+++ packages/ui-sdk/src/components/Charts/BubbleChart/index.tsx
@@ -28,9 +28,9 @@ export const BubbleChart: React.FC<IBubbleChartProps> = ({
   width,
   customTooltip,
 }) => {
-  const [_containerId] = useState(crypto.randomUUID());
-  const [_labelId] = useState(crypto.randomUUID());
-  const [_valueId] = useState(crypto.randomUUID());
+  const [_containerId] = useState(Math.random().toString(36));
+  const [_labelId] = useState(Math.random().toString(36));
+  const [_valueId] = useState(Math.random().toString(36));
   const chartRef = useRef<ChartJS>(null);
   const [chartData, setChartData] = useState<ChartData<"bubble">>({
     datasets: [],