+++ packages/ui-sdk/src/components/Charts/RadarChart/index.tsx
@@ -23,9 +23,9 @@ export const RadarChart: React.FC<IRadarChartProps> = ({
   width,
   customTooltip,
 }) => {
-  const [_containerId] = useState(crypto.randomUUID());
-  const [_labelId] = useState(crypto.randomUUID());
-  const [_valueId] = useState(crypto.randomUUID());
+  const [_containerId] = useState(Math.random().toString(36));
+  const [_labelId] = useState(Math.random().toString(36));
+  const [_valueId] = useState(Math.random().toString(36));
   const chartRef = useRef<ChartJS>(null);
   const [chartData, setChartData] = useState<ChartData<"radar">>({
     datasets: [],